anuvaad-etl/anuvaad-extractor/file-converter/src/requirements.txt
pip==20.1
nltk==3.5
kafka-python==2.0.1
flask==1.1.2
flask-cors==3.0.8
flask-restful==0.3.8
anuvaad-auditor==0.1.3
mongoengine==0.18.0
python-magic==0.4.22
